requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
timm>=0.9.0
albumentations>=1.3.0
